Organize project into packages
